+++ xmindparser/zenreader.py
@@ -1,7 +1,7 @@
 import json
 from zipfile import ZipFile
 
-from . import config, cache
+from . import cache, config
 
 content_json = "content.json"
 
@@ -13,7 +13,7 @@ def open_xmind(file_path):
         for f in xmind.namelist():
             for key in [content_json]:
                 if f == key:
-                    cache[key] = xmind.open(f).read().decode('utf-8')
+                    cache[key] = xmind.open(f).read().decode("utf-8")
 
 
 def get_sheets():
@@ -24,99 +24,120 @@ def get_sheets():
 
 def sheet_to_dict(sheet):
     """convert a sheet to dict type."""
-    topic = sheet['rootTopic']
-    result = {'title': sheet['title'], 'topic': node_to_dict(topic), 'structure': get_sheet_structure(sheet)}
+    topic = sheet["rootTopic"]
+    result = {"title": sheet["title"], "topic": node_to_dict(topic)}
 
-    if config['showTopicId']:
-        result['id'] = sheet['id']
+    if config["showTopicId"]:
+        result["id"] = sheet["id"]
 
-    if config['hideEmptyValue']:
+    if config["showStructure"]:
+        result["structure"] = get_sheet_structure(sheet)
+
+    if config["hideEmptyValue"]:
         result = {k: v for k, v in result.items() if v}
+
+    if config["showRelationship"]:
+        result["relationships"] = get_sheet_relationship(sheet)
 
     return result
 
 
 def get_sheet_structure(sheet):
-    root_topic = sheet['rootTopic']
-    return root_topic.get('structureClass', None)
+    root_topic = sheet["rootTopic"]
+    return root_topic.get("structureClass", None)
+
+
+def get_sheet_relationship(sheet):
+    if sheet.get("relationships", None) is None:
+        return []
+    relationship = sheet["relationships"]
+    for r in relationship:
+        if r.get("controlPoints", None) is not None:
+            del [r["controlPoints"]]
+        if r.get("titleUnedited", None) is not None:
+            del [r["titleUnedited"]]
+        if not config["showTopicId"]:
+            del [r["id"]]
+    return relationship
 
 
 def node_to_dict(node):
     """parse Element to dict data type."""
     child = children_topics_of(node)
 
-    d = {'title': node.get('title', ''),
-         'note': note_of(node),
-         'makers': maker_of(node),
-         'labels': labels_of(node),
-         'link': link_of(node),
-         'image': image_of(node),
-         'callout': callout_of(node)}
+    d = {
+        "title": node.get("title", ""),
+        "note": note_of(node),
+        "makers": maker_of(node),
+        "labels": labels_of(node),
+        "link": link_of(node),
+        "image": image_of(node),
+        "callout": callout_of(node),
+    }
 
-    if d['link']:
+    if d["link"]:
+        """if d['link'].startswith('xmind'):
+            d['link'] = '[To another xmind topic!]""" ""
 
-        if d['link'].startswith('xmind'):
-            d['link'] = '[To another xmind topic!]'
+        if d["link"].startswith("xap:attachments"):
+            del d["link"]
+            d["title"] = "[Attachment]{0}".format(d["title"])
 
-        if d['link'].startswith('xap:attachments'):
-            del d['link']
-            d['title'] = '[Attachment]{0}'.format(d['title'])
-
-    if child is not None:
-        d['topics'] = []
+    if child:
+        d["topics"] = []
         for c in child:
-            d['topics'].append(node_to_dict(c))
+            d["topics"].append(node_to_dict(c))
 
-    if config['showTopicId']:
-        d['id'] = node['id']
+    if config["showTopicId"]:
+        d["id"] = node["id"]
 
-    if config['hideEmptyValue']:
-        d = {k: v for k, v in d.items() if v or k == 'title'}
+    if config["hideEmptyValue"]:
+        d = {k: v for k, v in d.items() if v or k == "title"}
 
     return d
 
 
 def children_topics_of(topic_node):
-    children = topic_node.get('children', None)
+    children = topic_node.get("children", None)
 
     if children:
-        return children.get('attached', None)
+        return children.get("attached", None)
 
 
 def link_of(node):
-    return node.get('href', None)
+    return node.get("href", None)
 
 
 def image_of(node):
-    return node.get('image', None)
+    return node.get("image", None)
 
 
 def labels_of(node):
-    return node.get('labels', None)
+    return node.get("labels", None)
 
 
 def note_of(node):
-    note_node = node.get('notes', None)
+    note_node = node.get("notes", None)
 
     if note_node:
-        note = note_node.get('plain', None)
+        note = note_node.get("plain", None)
         if note:
-            return note.get('content', '').strip()
+            return note.get("content", "").strip()
 
 
 def maker_of(topic_node):
-    maker_node = topic_node.get('markers', None)
+    maker_node = topic_node.get("markers", None)
     if maker_node is not None:
         makers = []
         for maker in maker_node:
-            makers.append(maker.get('markerId', None))
+            makers.append(maker.get("markerId", None))
 
         return makers
 
 
 def callout_of(topic_node):
-    callout = topic_node.get('children', None)
+    callout = topic_node.get("children", None)
     if callout:
-        callout = callout.get('callout', None)
+        callout = callout.get("callout", None)
         if callout:
-            return [x['title'] for x in callout]
+            return [x["title"] for x in callout]